created another way of adding choices for question
previous method was inefficient way of adding choice objects to the
system It’d be better if we can add a bunch of Choices directly when you create the Question object.
ChoiceInline class will by default generate extraa choice objects and
it work like this There are three slots for related Choices – as specified by extra – and each time
you come back to the “Change” page for an already-created object, you get another three extra slots.